The Bambara village reaches the Niger: the walkable region grows a lobe to the
water instead of ending as a circle, the river is real geometry in the
settlement scene with a current the eye can read, and the adults' errands send
the villagers along it — around the building fabric, on a route, so they arrive
where they were sent.

+++ src/scenes/place/boundary.test.ts
@@ -28,7 +28,9 @@ describe('the walkable boundary', () => {
       expect(lut.length).toBe(BOUNDARY_LUT_SIZE)
       for (let j = 0; j < lut.length; j++) {
         const angle = ((j + 0.5) / lut.length) * Math.PI * 2
-        expect(lut[j]).toBe(placeBoundaryRadius(layout, angle))
+        // The lookup is a Float32Array, so the stored value is the single
+        // precision of the boundary — the same number, not a different one.
+        expect(lut[j]).toBe(Math.fround(placeBoundaryRadius(layout, angle)))
         expect(lut[j]).toBeGreaterThan(0)
       }
     }
